requirements.txt
Flask==3.0.2
Flask-MySQL==1.5.2
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
marshmallow==3.20.2
python-dotenv==1.0.1
gunicorn==21.2.0


